requirements.txt
tqdm
tabulate
pandas
yfinance
datetime
tk
matplotlib
pillow
